refactor(spec): extract builder-data and split temporal scale tests
Move authoring data snapshot/portable ISO conversion out of builder.ts into
builder-data.ts so GGBuilder sugar no longer shares a file with Date
materialization. Split the long temporal-api suite into schema vs authoring
tests and keep public builder re-exports stable.
